--- rules-tests/CodeQuality/Rector/Class_/AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector/Fixture/ManyToMany/getter_with_return_type.php.inc
@@ -1,54 +1,0 @@
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector\Fixture\ManyToMany;
-
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training;
-
-/**
- * @ORM\Entity
- */
-final class AdderParam
-{
-    /**
-     * @ORM\ManyToMany(targetEntity=Training::class, mappedBy="trainer")
-     */
-    private $trainings;
-
-    public function getTrainings(): Collection
-    {
-        return $this->trainings;
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector\Fixture\ManyToMany;
-
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training;
-
-/**
- * @ORM\Entity
- */
-final class AdderParam
-{
-    /**
-     * @ORM\ManyToMany(targetEntity=Training::class, mappedBy="trainer")
-     */
-    private $trainings;
-
-    /**
-     * @return \Doctrine\Common\Collections\Collection<int, \Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training>
-     */
-    public function getTrainings(): Collection
-    {
-        return $this->trainings;
-    }
-}
-
-?>
