--- rules-tests/CodeQuality/Rector/Class_/AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector/Fixture/OneToMany/not_a_collection_property.php.inc
@@ -1,45 +1,0 @@
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector\Fixture\OneToMany;
-
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training;
-
-/**
- * @ORM\Entity
- */
-final class AdderParam
-{
-    private $trainings;
-
-    public function getTrainings(): Collection
-    {
-        return $this->trainings;
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector\Fixture\OneToMany;
-
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training;
-
-/**
- * @ORM\Entity
- */
-final class AdderParam
-{
-    private $trainings;
-
-    public function getTrainings(): Collection
-    {
-        return $this->trainings;
-    }
-}
-
-?>
