--- rules-tests/CodeQuality/Rector/Class_/AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\Doctrine\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAnnotationRector::class);
-};
