--- rules-tests/CodeQuality/Rector/Class_/AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAttributeRector/Fixture/OneToMany/getter_with_return_type.php.inc
@@ -1,46 +1,0 @@
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAttributeRector\Fixture\OneToMany;
-
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training;
-
-#[ORM\Entity]
-final class AdderParam
-{
-    #[ORM\OneToMany(targetEntity:Training::class, mappedBy:"trainer")]
-    private $trainings;
-
-    public function getTrainings(): Collection
-    {
-        return $this->trainings;
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\AddReturnDocTypeToMethodWhichReturnsCollectionPropertyByAttributeRector\Fixture\OneToMany;
-
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training;
-
-#[ORM\Entity]
-final class AdderParam
-{
-    #[ORM\OneToMany(targetEntity:Training::class, mappedBy:"trainer")]
-    private $trainings;
-
-    /**
-     * @return \Doctrine\Common\Collections\Collection<int, \Rector\Doctrine\Tests\CodeQuality\Rector\Property\ImproveDoctrineCollectionDocTypeInEntityRector\Source\Training>
-     */
-    public function getTrainings(): Collection
-    {
-        return $this->trainings;
-    }
-}
-
-?>
